fixed bugs

+++ src/service/request.js
@@ -24,23 +24,23 @@ const processQueue = (error, token = null) => {
   failedQueue = [];
 };
 
-const refreshAccessToken = async refreshToken => {
+const refreshAccessToken = async (refreshToken, userId) => {
   try {
     const response = await fetch(`${SERVER_URL}/auth/refreshToken`, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify({refreshToken}),
+      body: JSON.stringify({
+        refreshToken,
+        id: userId,
+      }),
     });
 
-    if (!response.ok) {
-      throw new Error('Failed to refresh token');
-    }
-
     const res = await response.json();
-    const newAccessToken = res?.data?.accessToken;
-    await AsyncStorage.setItem('userToken', newAccessToken);
+    const newAccessToken = res?.data?.token;
+   
+    if (newAccessToken) await AsyncStorage.setItem('userToken', newAccessToken);
 
     return newAccessToken;
   } catch (error) {
@@ -69,6 +69,9 @@ request.addAsyncRequestTransform(async request => {
     const res = await response.json();
     const newToken = res?.data?.token;
     await AsyncStorage.setItem('userToken', newToken);
+    await AsyncStorage.setItem('refreshToken', res?.data?.refreshToken);
+    await AsyncStorage.setItem('userId', res?.data?._id);
+
     request.headers['jwttoken'] = `Bearer ${newToken}`;
   }
 });
@@ -82,49 +85,57 @@ request.axiosInstance.interceptors.response.use(
       error.response &&
       (error.response.status === 401 || error.response.status === 403)
     ) {
-      if (!isRefreshing) {
-        isRefreshing = true;
+      if (!originalRequest._retry) {
+        originalRequest._retry = true;
 
-        try {
-          const refreshToken = await AsyncStorage.getItem('refreshToken');
-          if (!refreshToken) {
-            throw new Error('No refresh token available');
+        if (!isRefreshing) {
+          isRefreshing = true;
+
+          try {
+            const refreshToken = await AsyncStorage.getItem('refreshToken');
+            const userId = await AsyncStorage.getItem('userId');
+
+            if (!refreshToken) {
+              throw new Error('No refresh token available');
+            }
+
+            const newAccessToken = await refreshAccessToken(
+              refreshToken,
+              userId,
+            );
+            processQueue(null, newAccessToken);
+
+            originalRequest.headers['jwttoken'] = `Bearer ${newAccessToken}`;
+            return request.axiosInstance(originalRequest);
+          } catch (refreshError) {
+            processQueue(refreshError, null);
+
+            if (
+              refreshError.message
+                .toLowerCase()
+                .includes('invalid refreshtoken')
+            ) {
+              await AsyncStorage.clear(); // or any other action to log out the user
+              console.error('Refresh token is invalid, logging out the user.');
+            }
+
+            return Promise.reject(refreshError);
+          } finally {
+            isRefreshing = false;
           }
-
-          const newAccessToken = await refreshAccessToken(refreshToken);
-          processQueue(null, newAccessToken);
-
-          isRefreshing = false;
-
-          originalRequest.headers['jwttoken'] = `Bearer ${newAccessToken}`;
-          return request.axiosInstance(originalRequest);
-        } catch (refreshError) {
-          processQueue(refreshError, null);
-
-          if (
-            refreshError.message.toLowerCase().includes('invalid refreshtoken')
-          ) {
-            // Handle the case where the refresh token is invalid
-            await AsyncStorage.clear(); // or any other action to log out the user
-            console.error('Refresh token is invalid, logging out the user.');
-          }
-
-          return Promise.reject(refreshError);
-        } finally {
-          isRefreshing = false;
         }
-      }
 
-      return new Promise((resolve, reject) => {
-        failedQueue.push({resolve, reject});
-      })
-        .then(token => {
-          originalRequest.headers['jwttoken'] = `Bearer ${token}`;
-          return request.axiosInstance(originalRequest);
+        return new Promise((resolve, reject) => {
+          failedQueue.push({resolve, reject});
         })
-        .catch(err => {
-          return Promise.reject(err);
-        });
+          .then(token => {
+            originalRequest.headers['jwttoken'] = `Bearer ${token}`;
+            return request.axiosInstance(originalRequest);
+          })
+          .catch(err => {
+            return Promise.reject(err);
+          });
+      }
     }
 
     return Promise.reject(error);
